+++ src/__tests__/foo.test.ts
@@ -1,4 +1,4 @@
-import { sum } from '../src/foo';
+import { sum } from '../foo';
 
 test('basic', () => {
   expect(sum()).toBe(0);
