+++ src/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { Counter } from '../src/index';
+import { Counter } from '../index';
 
 test('test', () => {
   const arr1_100 = Array.from(new Array(100)).map((v,i)=> i + 1);
